.config/fabric/requirements.txt
psutil
git+https://github.com/Fabric-Development/fabric
python-dotenv
